+++ GenerateHTMLSummary.py
@@ -5,7 +5,7 @@ from Gui import *
 from collections import defaultdict
 
 # The script version. You can check the changelog at the GitHub URL to see if there is a new version.
-VERSION = "1.6.0"
+VERSION = "1.6.1"
 GITHUB_URL = "https://github.com/mbektic/Simple-SESH-Sumary/blob/main/CHANGELOG.md"