+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.1] 2025-04-29
+### Changed
+- Search now persists when switching between tabs. 
+
+### Fixed
+- Fixed mode toggle not working anymore after the tab update.
+
 ## [1.6.0] 2025-04-29
 ### Added
 - Year selector for the tables so you can see the data based off of the year.